+++ lib/servus/config.rb
@@ -9,11 +9,7 @@ module Servus
 
     def initialize
       # Default to Rails.root if available, otherwise use current working directory
-      @schema_root = if defined?(Rails)
-                       Rails.root.join('app/schemas/services')
-                     else
-                       File.expand_path('../../../app/schemas/services', __dir__)
-                     end
+      @schema_root = File.join(root_path, 'app/schemas/services')
     end
 
     # Returns the path for a specific service's schema
@@ -31,6 +27,19 @@ module Servus
     # @return [String] the directory for the service's schemas
     def schema_dir_for(service_namespace)
       File.join(schema_root.to_s, service_namespace)
+    end
+
+    private
+
+    # Sets the schema root directory
+    #
+    # @param path [String] the new schema root directory
+    def root_path
+      if defined?(Rails) && Rails.respond_to?(:root)
+        Rails.root
+      else
+        File.expand_path('../../..', __dir__)
+      end
     end
   end