+++ lib/servus/extensions/async/ext.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module Servus
+  module Extensions
+    module Async
+      require 'servus/extensions/async/errors'
+      require 'servus/extensions/async/job'
+      require 'servus/extensions/async/call'
+
+      # Module providing async extensions for Servus
+      module Ext; end
+    end
+  end
+end
+